Connecting UpdateDataFisikActivity with KesehatanFisikActivity and Updating GET request to endpoint: updateData on Android

+++ Android/app/src/main/java/com/example/medicare/data/network/ApiService.kt
@@ -16,7 +16,7 @@ interface ApiService {
             @Query("glukosa") glukosa : Double,
             @Query("jumlahKehamilan") jumlahKehamilan : Double,
             @Query("ketebalanKulit") ketebalanKulit : Double,
-            @Query("insulin") insulin : Int,
+            @Query("insulin") insulin : Double,
             @Query("age") age : Int
     ) : Call<DiabetesPrediction>
 }
